+++ packages/compiler-cli/src/ngtsc/annotations/directive/index.ts
@@ -9,3 +9,4 @@
 export {DirectiveDecoratorHandler} from './src/handler';
 export {DirectiveSymbol} from './src/symbol';
 export * from './src/shared';
+export * from './src/input_function';
